refactor & transferring code to a component
